VRAM::Delegatable, Terminus::Drawable and Shinonome::Drawable
This patch moves some duplicated methods to dedicated modules:
- VRAM::Delegatable ... methods like `draw_bitmap`
- Terminus::Drawable ... draw_terminus
- Shinonome::Drawable ... draw_shononome

UC8151 and SSD1306 use these modules
