+++ mrbgems/picoruby-ssd1306/sig/ssd1306.rbs
@@ -31,14 +31,10 @@ class SSD1306
   def clear: () -> Integer
   def erase: (Integer x, Integer y, Integer w, Integer h) -> nil
   def fill_screen: (?Integer pattern) -> Integer
-  def set_pixel: (Integer x, Integer y, ?Integer color) -> void
-  def draw_line: (Integer x0, Integer y0, Integer x1, Integer y1, ?Integer color) -> nil
-  def draw_rect: (Integer x0, Integer y0, Integer x1, Integer y1, ?Integer color, ?bool fill) -> nil
-  def draw_bitmap: (x: Integer, y: Integer, w: Integer, h: Integer, data: Array[Integer]) -> nil
-  def draw_bytes: (x: Integer, y: Integer, w: Integer, h: Integer, data: String) -> nil
+  include VRAM::Delegatable
+  include Terminus::Drawable
+  include Shinonome::Drawable
   def draw_text: (Symbol | String fontname, Integer x, Integer y, String text, ?Integer scale) -> nil
-  def draw_shinonome: (String size, Integer x, Integer y, String text, ?Integer scale) -> nil
-  def draw_terminus: (String size, Integer x, Integer y, String text, ?Integer scale) -> nil
   def update_display: () -> Integer
   def update_display_optimized: () -> Integer # Returns number of pages updated
 end
