+++ mrbgems/picoruby-vram/sig/vram.rbs
@@ -14,4 +14,15 @@ class VRAM
   def draw_bytes: (x: Integer, y: Integer, w: Integer, h: Integer, data: String) -> self
   def fill: (Integer color) -> self
   def erase: (Integer x, Integer y, Integer w, Integer h) -> self
+
+  module Delegatable
+    @width: Integer
+    @height: Integer
+    @vram: VRAM
+    def set_pixel: (Integer x, Integer y, ?Integer color) -> void
+    def draw_bitmap: (x: Integer, y: Integer, w: Integer, h: Integer, data: Array[Integer]) -> nil
+    def draw_bytes: (x: Integer, y: Integer, w: Integer, h: Integer, data: String) -> nil
+    def draw_line: (Integer x0, Integer y0, Integer x1, Integer y1, ?Integer color) -> nil
+    def draw_rect: (Integer x, Integer y, Integer w, Integer h, ?Integer color, ?bool fill) -> nil
+  end
 end
